+++ src/rex.mart/demo/src/rex/mart_demo.py
@@ -13,7 +13,7 @@ class DemoMartAccessPermissions(MartAccessPermissions):
     def user_can_access_definition(cls, user, definition_or_id):
         if user == 'test':
             return True
-        allowed = ('some_data', 'some_more_data', 'empty', 'broken_sql')
+        allowed = ('some_data', 'some_more_data', 'empty', 'broken_sql', 'some_parameters')
         if isinstance(definition_or_id, basestring):
             return definition_or_id in allowed
         else: