+++ src/rex.mart/demo/static/params2_etl.sql
@@ -1,0 +1,3 @@
+insert into foo (col1) values (CONCAT(%(prefix)s, %(DEFINITION)s));
+insert into foo (col1) values (CONCAT(%(prefix)s, %(foo)s));
+insert into foo (col1) values (CONCAT(%(prefix)s, %(bar)s));